move out the generator to a separate file
